+++ crates/core/src/ecc/signers/eip191/mod.rs
@@ -31,10 +31,7 @@ pub fn recover(msg: &[u8], sig: impl AsRef<[u8]>) -> Result<PublicKey<33>> {
     let sig_byte: [u8; 65] = sig.as_ref().try_into()?;
     let hash = hash(msg);
     let mut sig712 = sig_byte;
-    if !(27..=30).contains(&sig712[64]) {
-        return Err(crate::error::Error::InvalidRecoverId(sig712[64]));
-    }
-    sig712[64] -= 27;
+    sig712[64] = super::recovery_id_from_v(sig712[64], 27)?;
     crate::ecc::recover_hash(&hash, &sig712)
 }